requirements.txt
amqp==5.0.6
appdirs==1.4.4
asgiref==3.3.4
beautifulsoup4==4.9.3
billiard==3.6.4.0
celery==5.0.5
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
distlib==0.3.1
Django==3.2
django-ckeditor==6.1.0
django-common-helpers==0.9.2
django-cors-headers==3.7.0
django-debug-toolbar==3.2.1
django-etc==1.3.0
django-filter==2.4.0
django-hitcount==1.3.3
django-js-asset==1.2.2
django-phonenumber-field==5.1.0
django-resized==0.3.11
django-rest-serializer-field-permissions==4.0.0
django-taggit==1.4.0
django-tinymce==3.3.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.0
drf-yasg==1.20.0
filelock==3.0.12
future==0.18.2
google==3.0.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
kombu==5.1.0
MarkupSafe==2.0.1
oauthlib==3.1.1
openapi-codec==1.3.2
packaging==20.9
phonenumbers==8.12.23
phonenumberslite==8.12.22
Pillow==8.2.0
prompt-toolkit==3.0.18
pycparser==2.20
PyJWT==2.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.18.0
python-social-auth==0.3.6
python3-openid==3.2.0
pytils==0.3
pytz==2021.1
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
restframework-serializer-permissions==0.0.2
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
serializers==0.2.4
simplejson==3.17.2
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
soupsieve==2.2.1
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.4
vine==5.0.0
virtualenv==20.4.6
wcwidth==0.2.5
